@@ -1,9 +1,13 @@
 {
   "name": "react-dom",
-  "version": "16.5.2",
+  "version": "18.2.0",
   "description": "React package for working with the DOM.",
   "main": "index.js",
-  "repository": "facebook/react",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/react.git",
+    "directory": "packages/react-dom"
+  },
   "keywords": [
     "react"
   ],
@@ -14,26 +18,39 @@
   "homepage": "https://reactjs.org/",
   "dependencies": {
     "loose-envify": "^1.1.0",
-    "object-assign": "^4.1.1",
-    "prop-types": "^15.6.2",
-    "schedule": "^0.5.0"
+    "scheduler": "^0.23.0"
   },
   "peerDependencies": {
-    "react": "^16.0.0"
+    "react": "^18.2.0"
   },
   "files": [
     "LICENSE",
     "README.md",
     "index.js",
+    "client.js",
     "profiling.js",
     "server.js",
     "server.browser.js",
     "server.node.js",
     "test-utils.js",
-    "unstable-native-dependencies.js",
     "cjs/",
     "umd/"
   ],
+  "exports": {
+    ".": "./index.js",
+    "./client": "./client.js",
+    "./server": {
+      "deno": "./server.browser.js",
+      "worker": "./server.browser.js",
+      "browser": "./server.browser.js",
+      "default": "./server.node.js"
+    },
+    "./server.browser": "./server.browser.js",
+    "./server.node": "./server.node.js",
+    "./profiling": "./profiling.js",
+    "./test-utils": "./test-utils.js",
+    "./package.json": "./package.json"
+  },
   "browser": {
     "./server.js": "./server.browser.js"
   },
