@@ -1,92 +1,7 @@
 {
   "name": "react-router-dom",
-  "version": "4.3.1",
-  "description": "DOM bindings for React Router",
-  "repository": "ReactTraining/react-router",
-  "license": "MIT",
-  "authors": [
-    "Michael Jackson",
-    "Ryan Florence"
-  ],
-  "files": [
-    "BrowserRouter.js",
-    "HashRouter.js",
-    "Link.js",
-    "MemoryRouter.js",
-    "NavLink.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "es",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "withRouter.js",
-    "umd"
-  ],
-  "main": "index.js",
-  "module": "es/index.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "node ./tools/build.js",
-    "watch": "babel ./modules -d . --ignore __tests__ --watch",
-    "prepublishOnly": "node ./tools/build.js",
-    "clean": "git clean -fdX .",
-    "lint": "eslint modules",
-    "test": "jest"
-  },
-  "peerDependencies": {
-    "react": ">=15"
-  },
-  "dependencies": {
-    "history": "^4.7.2",
-    "invariant": "^2.2.4",
-    "loose-envify": "^1.3.1",
-    "prop-types": "^15.6.1",
-    "react-router": "^4.3.1",
-    "warning": "^4.0.1"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.2.3",
-    "babel-jest": "^23.0.1",
-    "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-imports": "^1.5.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "eslint": "^4.19.1",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-react": "^7.9.1",
-    "gzip-size": "^4.1.0",
-    "jest": "^23.1.0",
-    "pretty-bytes": "^5.0.0",
-    "raf": "^3.4.0",
-    "react": "^16.4.0",
-    "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^16.4.0",
-    "rollup": "^0.60.0",
-    "rollup-plugin-babel": "^3.0.4",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-uglify": "^3.0.0"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
-  "jest": {
-    "setupFiles": [
-      "raf/polyfill"
-    ]
-  },
+  "version": "6.22.0",
+  "description": "Declarative routing for React web applications",
   "keywords": [
     "react",
     "router",
@@ -94,5 +9,41 @@
     "routing",
     "history",
     "link"
-  ]
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/remix-run/react-router",
+    "directory": "packages/react-router-dom"
+  },
+  "license": "MIT",
+  "author": "Remix Software <hello@remix.run>",
+  "sideEffects": false,
+  "main": "./dist/main.js",
+  "unpkg": "./dist/umd/react-router-dom.production.min.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "dependencies": {
+    "@remix-run/router": "1.15.0",
+    "react-router": "6.22.0"
+  },
+  "devDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "peerDependencies": {
+    "react": ">=16.8",
+    "react-dom": ">=16.8"
+  },
+  "files": [
+    "dist/",
+    "CHANGELOG.md",
+    "LICENSE.md",
+    "README.md",
+    "server.d.ts",
+    "server.js",
+    "server.mjs"
+  ],
+  "engines": {
+    "node": ">=14.0.0"
+  }
 }
