@@ -4,9 +4,9 @@
   "private": true,
   "dependencies": {
     "antd": "^3.2.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-router-dom": "^4.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "1.1.5"
   },
   "scripts": {
@@ -16,8 +16,11 @@
     "eject": "react-scripts eject"
   },
   "devDependencies": {
+    "@nextui-org/react": "^2.2.9",
     "babel-plugin-import": "^1.6.5",
+    "framer-motion": "^11.0.5",
     "react-app-rewire-less": "^2.1.0",
-    "react-app-rewired": "^1.4.1"
+    "react-app-rewired": "^1.4.1",
+    "tailwindcss": "^3.4.1"
   }
 }
