@@ -1,25 +1,21 @@
-import React, { Component } from 'react';
-import './App.css';
-import {
-  Route,
-  withRouter,
-  Switch
-} from 'react-router-dom';
+import React, { Component } from "react";
+import "./App.css";
+import { Route, Routes, withRouter, BrowserRouter } from "react-router-dom";
 
-import { getCurrentUser } from '../util/APIUtils';
-import { ACCESS_TOKEN } from '../constants';
+import { getCurrentUser } from "../util/APIUtils";
+import { ACCESS_TOKEN } from "../constants";
 
-import TaskList from '../task/TaskList';
-import NewTask from '../task/NewTask';
-import Login from '../user/login/Login';
-import Signup from '../user/signup/Signup';
-import Profile from '../user/profile/Profile';
-import AppHeader from '../common/AppHeader';
-import NotFound from '../common/NotFound';
-import LoadingIndicator from '../common/LoadingIndicator';
-import PrivateRoute from '../common/PrivateRoute';
+import TaskList from "../task/TaskList";
+import NewTask from "../task/NewTask";
+import Login from "../user/login/Login";
+import Signup from "../user/signup/Signup";
+import Profile from "../user/profile/Profile";
+import AppHeader from "../common/AppHeader";
+import NotFound from "../common/NotFound";
+import LoadingIndicator from "../common/LoadingIndicator";
+import PrivateRoute from "../common/PrivateRoute";
 
-import { Layout, notification } from 'antd';
+import { Layout, notification } from "antd";
 const { Content } = Layout;
 
 class App extends Component {
@@ -28,57 +24,62 @@ class App extends Component {
     this.state = {
       currentUser: null,
       isAuthenticated: false,
-      isLoading: true
-    }
+      isLoading: true,
+    };
     this.handleLogout = this.handleLogout.bind(this);
     this.loadCurrentUser = this.loadCurrentUser.bind(this);
     this.handleLogin = this.handleLogin.bind(this);
 
     notification.config({
-      placement: 'topRight',
+      placement: "topRight",
       top: 70,
       duration: 3,
-    });    
+    });
   }
 
   loadCurrentUser() {
     getCurrentUser()
-    .then(response => {
-      this.setState({
-        currentUser: response,
-        isAuthenticated: true,
-        isLoading: false
+      .then((response) => {
+        this.setState({
+          currentUser: response,
+          isAuthenticated: true,
+          isLoading: false,
+        });
+      })
+      .catch((error) => {
+        this.setState({
+          isLoading: false,
+        });
       });
-    }).catch(error => {
-      this.setState({
-        isLoading: false
-      });  
-    });
   }
 
   componentDidMount() {
     this.loadCurrentUser();
   }
 
-  handleLogout(redirectTo="/", notificationType="success", description="You're successfully logged out.") {
+  handleLogout(
+    redirectTo = "/",
+    notificationType = "success",
+    description = "You're successfully logged out."
+  ) {
     localStorage.removeItem(ACCESS_TOKEN);
 
     this.setState({
       currentUser: null,
-      isAuthenticated: false
+      isAuthenticated: false,
     });
 
     this.props.history.push(redirectTo);
-    
+
     notification[notificationType]({
-      message: 'Tasking App',
+      message: "Tasking App",
       description: description,
     });
   }
 
   handleLogin() {
     notification.success({
-      message: 'Tasking App',
+      message: "Tasking App",
       description: "You're successfully logged in.",
     });
     this.loadCurrentUser();
@@ -86,35 +87,63 @@ class App extends Component {
   }
 
   render() {
-    if(this.state.isLoading) {
-      return <LoadingIndicator />
+    if (this.state.isLoading) {
+      return <LoadingIndicator />;
     }
-    
-    return (
-        <Layout className="app-container">
-          <AppHeader isAuthenticated={this.state.isAuthenticated} 
-            currentUser={this.state.currentUser} 
-            onLogout={this.handleLogout} />
 
-          <Content className="app-content">
-            <div className="container">
-              <Switch>      
-                <Route exact path="/" 
-                  render={(props) => <TaskList isAuthenticated={this.state.isAuthenticated} 
-                      currentUser={this.state.currentUser} handleLogout={this.handleLogout} {...props} />}>
-                </Route>
-                <Route path="/login" 
-                  render={(props) => <Login onLogin={this.handleLogin} {...props} />}></Route>
+    return (
+      <Layout className="app-container">
+        <AppHeader
+          isAuthenticated={this.state.isAuthenticated}
+          currentUser={this.state.currentUser}
+          onLogout={this.handleLogout}
+        />
+
+        <Content className="app-content">
+          <div className="container">
+            <BrowserRouter>
+              <Routes>
+                <Route
+                  exact
+                  path="/"
+                  render={(props) => (
+                    <TaskList
+                      isAuthenticated={this.state.isAuthenticated}
+                      currentUser={this.state.currentUser}
+                      handleLogout={this.handleLogout}
+                      {...props}
+                    />
+                  )}
+                ></Route>
+                <Route
+                  path="/login"
+                  render={(props) => (
+                    <Login onLogin={this.handleLogin} {...props} />
+                  )}
+                ></Route>
                 <Route path="/signup" component={Signup}></Route>
-                <Route path="/users/:username" 
-                  render={(props) => <Profile isAuthenticated={this.state.isAuthenticated} currentUser={this.state.currentUser} {...props}  />}>
-                </Route>
-                <PrivateRoute authenticated={this.state.isAuthenticated} path="/task/new" component={NewTask} handleLogout={this.handleLogout}></PrivateRoute>
+                <Route
+                  path="/users/:username"
+                  render={(props) => (
+                    <Profile
+                      isAuthenticated={this.state.isAuthenticated}
+                      currentUser={this.state.currentUser}
+                      {...props}
+                    />
+                  )}
+                ></Route>
+                <PrivateRoute
+                  authenticated={this.state.isAuthenticated}
+                  path="/task/new"
+                  component={NewTask}
+                  handleLogout={this.handleLogout}
+                ></PrivateRoute>
                 <Route component={NotFound}></Route>
-              </Switch>
-            </div>
-          </Content>
-        </Layout>
+              </Routes>
+            </BrowserRouter>
+          </div>
+        </Content>
+      </Layout>
     );
   }
 }
