@@ -1,73 +1,73 @@
 import React, { Component } from 'react';
 import {
-    Link,
-    withRouter
+  Link,
+  withRouter
 } from 'react-router-dom';
 import './AppHeader.css';
 import taskIcon from '../task.svg';
 import { Layout, Menu, Dropdown, Icon } from 'antd';
 const Header = Layout.Header;
-    
+
 class AppHeader extends Component {
-    constructor(props) {
-        super(props);   
-        this.handleMenuClick = this.handleMenuClick.bind(this);   
+  constructor(props) {
+    super(props);
+    this.handleMenuClick = this.handleMenuClick.bind(this);
+  }
+
+  handleMenuClick({ key }) {
+    if (key === "logout") {
+      this.props.onLogout();
+    }
+  }
+
+  render() {
+    let menuItems;
+    if (this.props.currentUser) {
+      menuItems = [
+        <Menu.Item key="/">
+          <Link to="/">
+            <Icon type="home" className="nav-icon" />
+          </Link>
+        </Menu.Item>,
+        <Menu.Item key="/task/new">
+          <Link to="/task/new">
+            <img src={taskIcon} alt="task" className="task-icon" />
+          </Link>
+        </Menu.Item>,
+        <Menu.Item key="/profile" className="profile-menu">
+          <ProfileDropdownMenu
+            currentUser={this.props.currentUser}
+            handleMenuClick={this.handleMenuClick} />
+        </Menu.Item>
+      ];
+    } else {
+      menuItems = [
+        <Menu.Item key="/login">
+          <Link to="/login">Login</Link>
+        </Menu.Item>,
+        <Menu.Item key="/signup">
+          <Link to="/signup">Signup</Link>
+        </Menu.Item>
+      ];
     }
 
-    handleMenuClick({ key }) {
-      if(key === "logout") {
-        this.props.onLogout();
-      }
-    }
-
-    render() {
-        let menuItems;
-        if(this.props.currentUser) {
-          menuItems = [
-            <Menu.Item key="/">
-              <Link to="/">
-                <Icon type="home" className="nav-icon" />
-              </Link>
-            </Menu.Item>,
-            <Menu.Item key="/task/new">
-            <Link to="/task/new">
-              <img src={taskIcon} alt="task" className="task-icon" />
-            </Link>
-          </Menu.Item>,
-          <Menu.Item key="/profile" className="profile-menu">
-                <ProfileDropdownMenu 
-                  currentUser={this.props.currentUser} 
-                  handleMenuClick={this.handleMenuClick}/>
-            </Menu.Item>
-          ]; 
-        } else {
-          menuItems = [
-            <Menu.Item key="/login">
-              <Link to="/login">Login</Link>
-            </Menu.Item>,
-            <Menu.Item key="/signup">
-              <Link to="/signup">Signup</Link>
-            </Menu.Item>                  
-          ];
-        }
-
-        return (
-            <Header className="app-header">
-            <div className="container">
-              <div className="app-title" >
-                <Link to="/">Tasking App</Link>
-              </div>
-              <Menu
-                className="app-menu"
-                mode="horizontal"
-                selectedKeys={[this.props.location.pathname]}
-                style={{ lineHeight: '64px' }} >
-                  {menuItems}
-              </Menu>
-            </div>
-          </Header>
-        );
-    }
+    return (
+      <Header className="app-header">
+        <div className="container">
+          <div className="app-title" >
+            <Link to="/">Todo App</Link>
+          </div>
+          <Menu
+            className="app-menu"
+            mode="horizontal"
+            selectedKeys={[this.props.location.pathname]}
+            style={{ lineHeight: '64px' }} >
+            {menuItems}
+          </Menu>
+        </div>
+      </Header>
+    );
+  }
 }
 
 function ProfileDropdownMenu(props) {
@@ -92,12 +92,12 @@ function ProfileDropdownMenu(props) {
   );
 
   return (
-    <Dropdown 
-      overlay={dropdownMenu} 
+    <Dropdown
+      overlay={dropdownMenu}
       trigger={['click']}
-      getPopupContainer = { () => document.getElementsByClassName('profile-menu')[0]}>
+      getPopupContainer={() => document.getElementsByClassName('profile-menu')[0]}>
       <a className="ant-dropdown-link">
-         <Icon type="user" className="nav-icon" style={{marginRight: 0}} /> <Icon type="down" />
+        <Icon type="user" className="nav-icon" style={{ marginRight: 0 }} /> <Icon type="down" />
       </a>
     </Dropdown>
   );
