@@ -4,12 +4,16 @@ import './index.css';
 import App from './app/App';
 import registerServiceWorker from './registerServiceWorker';
 import { BrowserRouter as Router } from 'react-router-dom';
+import { NextUIProvider } from '@nextui-org/react';
 
 ReactDOM.render(
-    <Router>
-        <App />
-    </Router>, 
-    document.getElementById('root')
+  <Router>
+
+    <NextUIProvider>
+      <App />
+    </NextUIProvider>
+  </Router>,
+  document.getElementById('root')
 );
 
 registerServiceWorker();
